refactor: Adiciona FINAL a injeção das minhas dependências, adiciona o thymeleaf para renderização, e faz modificações no MissoesController, para utilizar ResponseEntity

+++ src/main/java/dev/java10x/CadastroDeNinjas/Missoes/MissoesController.java
@@ -1,5 +1,8 @@
 package dev.java10x.CadastroDeNinjas.Missoes;
 
+import org.apache.catalina.connector.Response;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -8,35 +11,56 @@ import java.util.List;
 @RequestMapping("/missoes")
 public class MissoesController {
 
-    private MissoesService missoesService;
+    private final MissoesService missoesService;
 
     public MissoesController(MissoesService missoesService) {
         this.missoesService = missoesService;
     }
 
     @PostMapping("/criar")
-    public MissoesDTO criarMissao(@RequestBody MissoesDTO missao) {
-        return missoesService.criarMissao(missao);
+    public ResponseEntity<String> criarMissao(@RequestBody MissoesDTO missao) {
+        MissoesDTO missaoDTO = missoesService.criarMissao(missao);
+        return ResponseEntity.status(HttpStatus.CREATED)
+        .body("Missão criada com sucesso: " + missaoDTO.getNomeMissao());
     }
 
     @GetMapping("/listar")
-    public List<MissoesDTO> listarMissoes() {
-        return missoesService.listarMissoes();
+    public ResponseEntity<List<MissoesDTO>> listarMissoes() {
+        List<MissoesDTO> missoes = missoesService.listarMissoes();
+        return ResponseEntity.ok(missoes);
     }
 
     @GetMapping("/listar/{id}")
-    public MissoesDTO listarMissoesPorId(@PathVariable Long id) {
-        return missoesService.listarMissaoPorId(id);
+    public ResponseEntity<?> listarMissoesPorId(@PathVariable Long id) {
+        MissoesDTO missao = missoesService.listarMissaoPorId(id);
+        if (missao != null) {
+            return ResponseEntity.ok(missao);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+            .body("Missão com Id " + id + " não existe nos nossos registros");
+        }
     }
 
     @PutMapping("/alterar/{id}")
-    public MissoesDTO alterarMissao(@PathVariable Long id, @RequestBody MissoesDTO missao) {
-        return missoesService.alterarNinjaPorId(id, missao);
+    public ResponseEntity<?> alterarMissao(@PathVariable Long id, @RequestBody MissoesDTO missaoAtualizada) {
+        MissoesDTO missao = missoesService.alterarNinjaPorId(id, missaoAtualizada);
+        if (missao != null) {
+            return ResponseEntity.ok(missao);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+            .body("Missão com Id " + id + " não existe nos nossos registros");
+        }
     }
 
     @DeleteMapping("/deletar/{id}")
-    public void deletarMissao(@PathVariable Long id) {
-        missoesService.removerMissaoPorId(id);
+    public ResponseEntity<String> deletarMissao(@PathVariable Long id) {
+        if (missoesService.listarMissaoPorId(id) != null) {
+            missoesService.removerMissaoPorId(id);
+            return ResponseEntity.ok("Missão com Id " + id + " deletado com sucesso");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+            .body("Missão com Id " + id + " não encontrado");
+        }
     }
 
 }
